FIX: scratches on the surface in byobject mode
When printing by object, not lifted when traveling to the second piece, which causes scratches on the surface of the first piece.
Modify the lift height in GCode.cpp to add a z_hop height to the highest print height to avoid scratches.

jira: STUDIO-14001
Change-Id: I97835fce5b82f405d600f5aa6ce74edae1e97b47
(cherry picked from commit 9282806196e455883b36b17391a404106c4b3c82)